+++ qiskit_aqt_provider/aqt_resource.py
@@ -10,11 +10,13 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+import typing
 import warnings
 from dataclasses import dataclass
 from typing import (
     TYPE_CHECKING,
     Any,
+    Generic,
     Optional,
     TypeVar,
     Union,
@@ -35,7 +37,7 @@ from typing_extensions import override
 
 from qiskit_aqt_provider import api_models, api_models_direct
 from qiskit_aqt_provider.aqt_job import AQTDirectAccessJob, AQTJob
-from qiskit_aqt_provider.aqt_options import AQTOptions
+from qiskit_aqt_provider.aqt_options import AQTDirectAccessOptions, AQTOptions
 from qiskit_aqt_provider.circuit_to_aqt import aqt_to_qiskit_circuit
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -76,20 +78,27 @@ def make_transpiler_target(target_cls: type[TargetT], num_qubits: int) -> Target
 
 _JobType = TypeVar("_JobType", AQTJob, AQTDirectAccessJob)
 
+_OptionsType = TypeVar("_OptionsType", bound=AQTOptions)
+"""Resource options model."""
 
-class _ResourceBase(Backend):
+
+class _ResourceBase(Generic[_OptionsType], Backend):
     """Common setup for AQT backends."""
 
-    def __init__(self, provider: "AQTProvider", name: str):
+    def __init__(self, provider: "AQTProvider", name: str, options_type: type[_OptionsType]):
         """Initialize the Qiskit backend.
 
         Args:
             provider: Qiskit provider that owns this backend.
             name: name of the backend.
+            options_type: options model. Must be default-initializable.
         """
         super().__init__(name=name, provider=provider)
 
         num_qubits = 20
+        self._target = make_transpiler_target(Target, num_qubits)
+        self._options = options_type()
+
         self._configuration = BackendConfiguration.from_dict(
             {
                 "backend_name": name,
@@ -103,7 +112,7 @@ class _ResourceBase(Backend):
                 "memory": True,
                 "n_qubits": num_qubits,
                 "conditional": False,
-                "max_shots": 200,
+                "max_shots": self._options.max_shots(),
                 "max_experiments": 1,
                 "open_pulse": False,
                 "gates": [
@@ -113,8 +122,6 @@ class _ResourceBase(Backend):
                 ],
             }
         )
-        self._target = make_transpiler_target(Target, num_qubits)
-        self._options = AQTOptions()
 
     def configuration(self) -> BackendConfiguration:
         """Legacy Qiskit backend configuration."""
@@ -133,10 +140,11 @@ class _ResourceBase(Backend):
     @classmethod
     def _default_options(cls) -> QiskitOptions:
         """Default backend options, in Qiskit format."""
-        return QiskitOptions()
+        options_type = typing.get_args(cls.__orig_bases__[0])[0]
+        return QiskitOptions(**options_type())
 
     @property
-    def options(self) -> AQTOptions:
+    def options(self) -> _OptionsType:
         """Configured backend options."""
         return self._options
 
@@ -187,7 +195,7 @@ class _ResourceBase(Backend):
         )
 
 
-class AQTResource(_ResourceBase):
+class AQTResource(_ResourceBase[AQTOptions]):
     """Qiskit backend for AQT cloud quantum computing resources.
 
     Use :meth:`AQTProvider.get_backend <qiskit_aqt_provider.aqt_provider.AQTProvider.get_backend>`
@@ -205,7 +213,11 @@ class AQTResource(_ResourceBase):
             provider: Qiskit provider that owns this backend.
             resource_id: description of resource to target.
         """
-        super().__init__(name=resource_id.resource_id, provider=provider)
+        super().__init__(
+            name=resource_id.resource_id,
+            provider=provider,
+            options_type=AQTOptions,
+        )
 
         self._http_client: httpx.Client = provider._http_client
         self.resource_id = resource_id
@@ -266,7 +278,7 @@ class AQTResource(_ResourceBase):
         return api_models.Response.model_validate(resp.json())
 
 
-class AQTDirectAccessResource(_ResourceBase):
+class AQTDirectAccessResource(_ResourceBase[AQTDirectAccessOptions]):
     """Qiskit backend for AQT direct-access quantum computing resources.
 
     Use :meth:`AQTProvider.get_direct_access_backend <qiskit_aqt_provider.aqt_provider.AQTProvider.get_direct_access_backend>`
@@ -284,9 +296,13 @@ class AQTDirectAccessResource(_ResourceBase):
             provider: Qiskit provider that owns the backend.
             base_url: URL of the direct-access interface.
         """
-        self._http_client = api_models.http_client(base_url=base_url, token=provider.access_token)
+        super().__init__(
+            provider=provider,
+            name="direct-access",
+            options_type=AQTDirectAccessOptions,
+        )
 
-        super().__init__(provider=provider, name="direct-access")
+        self._http_client = api_models.http_client(base_url=base_url, token=provider.access_token)
 
     def run(
         self, circuits: Union[QuantumCircuit, list[QuantumCircuit]], **options: Any